+++ SingleExperience/Services/ListaProdutoCompra/Models/ItemProdutoCompraModel.cs
@@ -10,7 +10,8 @@ namespace SingleExperience.Services.ListaProdutoCompra.Models
         public int CompraId { get; set; }
         public int ProdutoId { get; set; }
         public string Nome { get; set; }
-        public double PrecoUnitario { get; set; }
+        public decimal PrecoUnitario { get; set; }
         public int Qtde { get; set; }
+
     }
 }
